fix(db): precisely assign BigInteger only to telegram and message identifier fields

+++ src/db/models.py
@@ -38,8 +38,8 @@ class User(Base):
     last_manual_report_date: Mapped[Optional[date]] = mapped_column(Date, nullable=True)
     
     # Catalyst Settings
-    catalyst_threshold_minutes: Mapped[int] = mapped_column(BigInteger, default=60)
-    catalyst_interval_minutes: Mapped[int] = mapped_column(BigInteger, default=20)
+    catalyst_threshold_minutes: Mapped[int] = mapped_column(Integer, default=60)
+    catalyst_interval_minutes: Mapped[int] = mapped_column(Integer, default=20)
     target_channel_id: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
     report_config: Mapped[Optional[Any]] = mapped_column(JSON, nullable=True)
 
@@ -87,8 +87,8 @@ class Project(Base):
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), index=True)
     title: Mapped[str] = mapped_column(String)
     status: Mapped[str] = mapped_column(String, default="active") # 'active', 'paused', 'completed'
-    target_value: Mapped[int] = mapped_column(BigInteger, default=0)
-    current_value: Mapped[int] = mapped_column(BigInteger, default=0)
+    target_value: Mapped[int] = mapped_column(Integer, default=0)
+    current_value: Mapped[int] = mapped_column(Integer, default=0)
     unit: Mapped[Optional[str]] = mapped_column(String, default="minutes")
     next_action_text: Mapped[Optional[str]] = mapped_column(String, nullable=True) # E.g., "Read pointers chapter"
 
@@ -98,14 +98,14 @@ class Habit(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), index=True)
     title: Mapped[str] = mapped_column(String)
-    target_value: Mapped[int] = mapped_column(BigInteger, default=1)
-    current_value: Mapped[int] = mapped_column(BigInteger, default=0)
+    target_value: Mapped[int] = mapped_column(Integer, default=1)
+    current_value: Mapped[int] = mapped_column(Integer, default=0)
     type: Mapped[str] = mapped_column(String, default="counter") # 'counter' or 'boolean'
     unit: Mapped[Optional[str]] = mapped_column(String, nullable=True)
     last_reset_date: Mapped[Optional[date]] = mapped_column(Date, nullable=True)
     created_at: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True), server_default=func.now())
-    total_completions: Mapped[int] = mapped_column(BigInteger, default=0)
-    current_streak: Mapped[int] = mapped_column(BigInteger, default=0)
+    total_completions: Mapped[int] = mapped_column(Integer, default=0)
+    current_streak: Mapped[int] = mapped_column(Integer, default=0)
 
 class TimeLog(Base):
     """The Ledger: Tracks actual focused work blocks"""
@@ -115,8 +115,8 @@ class TimeLog(Base):
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), index=True)
     session_id: Mapped[Optional[int]] = mapped_column(ForeignKey("sessions.id"), nullable=True)
     project_id: Mapped[Optional[int]] = mapped_column(ForeignKey("projects.id"), nullable=True)
-    duration_minutes: Mapped[int] = mapped_column(BigInteger)
-    progress_amount: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
+    duration_minutes: Mapped[int] = mapped_column(Integer)
+    progress_amount: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
     progress_unit: Mapped[Optional[str]] = mapped_column(String, nullable=True)
     description: Mapped[Optional[str]] = mapped_column(String, nullable=True)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
@@ -147,8 +147,8 @@ class TokenUsage(Base):
     
     id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.id"), index=True)
-    prompt_tokens: Mapped[int] = mapped_column(BigInteger, default=0)
-    completion_tokens: Mapped[int] = mapped_column(BigInteger, default=0)
-    total_tokens: Mapped[int] = mapped_column(BigInteger, default=0)
+    prompt_tokens: Mapped[int] = mapped_column(Integer, default=0)
+    completion_tokens: Mapped[int] = mapped_column(Integer, default=0)
+    total_tokens: Mapped[int] = mapped_column(Integer, default=0)
     model_name: Mapped[str] = mapped_column(String)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
